VEOSS-1787 | Allow Excluded Traders to Re-Register For the OSS Scheme (#390)
* VEOSS-1787 | Allow Excluded Traders to Re-Register For the OSS Scheme

* VEOSS-1787 | Added Extra tests and RejoinJourneyRecoveryController

* VEOSS-1787 | Added missing tests for RejoinRegistrationController

* VEOSS-1787 | Added missing onSubmit tests for RejoinRegistrationController

* VEOSS-1787 | Remove Save and come back button and Date of first sale answers if rejoin

* VEOSS-1787 | Remove saved answers Have you made sales page on start of rejoin journey

* VEOSS-1787 | Fix issue with failure page redirecting to wrong view

* VEOSS-1787 | Added RejoinMode to pages to keep in rejoin mode when make changes to answers

* VEOSS-1787 | Fix issue with navigation to wrong page after adding previous registartion

* VEOSS-1787 | Fix issue found during testing

* VEOSS-1787 | Show change link for previous registrations on rejoin if it gets pulled back from ETMP with a no

* VEOSS-1787 | Fix navigation for when cghanging Registered for tax in other EU countries answer to No

* VEOSS-1787 | Fix issue with remove link showing on previous registartions

* VEOSS-1787 | Added Rejoin Mode to DateOfFirstSale and BusinessContactDetails

* VEOSS-1787 | Fix issue with invalid date on date of first sale page in rejoin mode

* VEOSS-1787 | Fix issue with invalid date on date of first sale page in rejoin mode

* VEOSS-1787 | Fix issue with manually navigating to delete-previous-schemes and navigation with remove websites

+++ app/controllers/previousRegistrations/DeletePreviousSchemeController.scala
@@ -20,12 +20,13 @@ import config.Constants.lastSchemeForCountry
 import controllers.GetCountry
 import controllers.actions._
 import forms.previousRegistrations.DeletePreviousSchemeFormProvider
-import models.{Index, Mode}
+import models.{AmendMode, Index, Mode, RejoinMode}
 import pages.previousRegistrations.DeletePreviousSchemePage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import queries.previousRegistration.{DeriveNumberOfPreviousSchemes, PreviousSchemeForCountryQuery}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
+import utils.CheckExistingRegistrations.getExistingRegistrationSchemes
 import viewmodels.checkAnswers.previousRegistrations.{DeletePreviousSchemeSummary, PreviousIntermediaryNumberSummary, PreviousSchemeNumberSummary}
 import viewmodels.govuk.summarylist._
 import views.html.previousRegistrations.DeletePreviousSchemeView
@@ -66,7 +67,16 @@ class DeletePreviousSchemeController @Inject()(
             case Some(value) => form.fill(value)
           }
 
-          Future.successful(Ok(view(preparedForm, mode, countryIndex, schemeIndex, country, list, isLastPreviousScheme)))
+          if (mode == RejoinMode || mode == AmendMode) {
+            val existingSchemes = getExistingRegistrationSchemes(country)
+            if (existingSchemes.nonEmpty) {
+              Future.successful(Redirect(routes.CannotRemoveExistingPreviousSchemesController.onPageLoad()))
+            } else {
+              Future.successful(Ok(view(preparedForm, mode, countryIndex, schemeIndex, country, list, isLastPreviousScheme)))
+            }
+          } else {
+            Future.successful(Ok(view(preparedForm, mode, countryIndex, schemeIndex, country, list, isLastPreviousScheme)))
+          }
       }
   }
 
@@ -100,7 +110,16 @@ class DeletePreviousSchemeController @Inject()(
                   _ <- cc.sessionRepository.set(updatedAnswers)
                 } yield Redirect(DeletePreviousSchemePage(countryIndex).navigate(mode, updatedAnswers))
               } else {
-                Future.successful(Redirect(DeletePreviousSchemePage(countryIndex).navigate(mode, request.userAnswers)))
+                if (mode == RejoinMode || mode == AmendMode) {
+                  val existingSchemes = getExistingRegistrationSchemes(country)
+                  if (existingSchemes.nonEmpty) {
+                    Future.successful(Redirect(routes.CannotRemoveExistingPreviousSchemesController.onPageLoad()))
+                  } else {
+                    Future.successful(Redirect(DeletePreviousSchemePage(countryIndex).navigate(mode, request.userAnswers)))
+                  }
+                } else {
+                  Future.successful(Redirect(DeletePreviousSchemePage(countryIndex).navigate(mode, request.userAnswers)))
+                }
               }
           )
       }
